gurumettoai/requirements.txt
Django==4.2.5
djangorestframework==3.15.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
openai==0.28.0
